+++ lib/potassium/templates/application/assets/config/environments/staging.rb
@@ -1,0 +1,5 @@
+require_relative "production"
+
+Rails.application.configure do
+  # ...
+end