+++ lib/potassium/templates/application/recipes/asks/heroku.rb
@@ -1,0 +1,3 @@
+heroku = Ask.confirm "Are you going to deploy to heroku?"
+
+set :heroku, heroku if heroku