+++ lib/potassium/templates/application/recipes/paperclip.rb
@@ -1,20 +1,17 @@
 if selected?(:paperclip)
   gather_gem('paperclip', '~> 4.3')
 
-  gsub_file "config/environments/production.rb", /^end$/o do |_match|
-    %{
-
-  # Paperclip support for S3
+  paperclip_config = %{
   config.paperclip_defaults = {
     :storage => :s3,
     :s3_credentials => {
       :bucket => ENV['AWS_BUCKET']
     }
   }
-end}
-  end
+  }
 
-  append_to_file '.rbenv-vars.example', 'AWS_BUCKET='
-  append_to_file '.rbenv-vars', 'AWS_BUCKET='
+  application paperclip_config.strip, env: 'production'
 
+  append_to_file '.env.example', 'AWS_BUCKET='
+  append_to_file '.env', 'AWS_BUCKET='
 end