+++ lib/potassium/templates/application/recipes/unicorn.rb
@@ -1,5 +1,0 @@
-gather_gem 'unicorn-rails'
-
-gather_gems(:production) do
-  gather_gem 'unicorn'
-end