+++ lib/potassium/templates/application/template.rb
@@ -2,12 +2,6 @@ set :app_name, @app_name
 set :titleized_app_name, get(:app_name).titleize
 set :underscorized_app_name, get(:app_name).underscore
 
-default_env(
-  'DB_NAME' => get(:underscorized_app_name).to_s,
-  'DB_USER' => "root",
-  'DB_PASSWORD' => ''
-)
-
 run_action(:cleaning) do
   clean_gemfile
   gather_gem('spring')
@@ -21,13 +15,16 @@ run_action(:asking) do
   eval_file "recipes/asks/i18n.rb"
   eval_file "recipes/asks/api.rb"
   eval_file "recipes/asks/paperclip.rb"
+  eval_file "recipes/asks/heroku.rb"
 end
 
 run_action(:recipe_loading) do
-  eval_file "recipes/unicorn.rb"
+  eval_file "recipes/heroku.rb"
+  eval_file "recipes/puma.rb"
   eval_file "recipes/database.rb"
   eval_file "recipes/readme.rb"
-  eval_file "recipes/rbenv.rb"
+  eval_file "recipes/ruby.rb"
+  eval_file "recipes/env.rb"
   eval_file "recipes/bower.rb"
   eval_file "recipes/editorconfig.rb"
   eval_file "recipes/aws_sdk.rb"
@@ -39,10 +36,13 @@ run_action(:recipe_loading) do
   eval_file "recipes/pundit.rb"
   eval_file "recipes/testing.rb"
   eval_file "recipes/production.rb"
+  eval_file "recipes/staging.rb"
+  eval_file "recipes/secrets.rb"
   eval_file "recipes/git.rb"
   eval_file "recipes/api.rb"
   eval_file "recipes/rack-cors.rb"
   eval_file "recipes/paperclip.rb"
+  eval_file "recipes/cleanup.rb"
 end
 
 say "Gathered enough information. Applying the template. Wait a minute.", :green